Harden async PowerShell pipeline lifecycle
